VSOCK Implementation for LibIIO
libiio: network context support for UNIX domain and virtio sockets

libiio has support for a network context object, which enables connecting
to an iiod server running on a remote machine via TCP/IP.

For Chrome OS purposes, we want to run iiod over a UNIX socket or a vsock
to enable either VMs or local processes to connect and manage sensors, but
not remote devices across the network.

This patch adds support to the server and the client sides of libiio/iiod
to create UNIX domain and virtio sockets, and establish connections over
such sockets.

reference:
https://chromium-review.googlesource.com/c/chromiumos/third_party/libiio/+/1719570

Tracked-On: OAM-100107
Signed-off-by: vilasrk <vilas.r.k@intel.com>