Changing Gyroscope Sensor type to uncalibrated.
In virtualization environment getting sensor data from
host has delay due to iio interface which is resulting
in CTS verifier failure to fix this we are changing
Gyroscope sensor type to uncalibrated.

Uncalibrated gyroscope, which is similar to the gyroscope
base sensor, but with the bias calibration being reported
separately instead of being corrected in the measurement.

Test:
All the CTSSensorTest cases passed.
All the CTSVerifier Test cases passed.

Tracked-On: OAM-102909
Signed-off-by: vilasrk vilas.r.k@intel.com
Signed-off-by: RajaniRanjan <rajani.ranjan@intel.com>